refactor: Application 계층과 Web 계층 간 DTO 작성
