+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/application/dto/CommentInfo.kt
@@ -1,0 +1,9 @@
+package bingsoochef.bingsoochef.toppping.application.dto
+
+import java.time.LocalDateTime
+
+data class CommentInfo (
+    val id: Long,
+    val content: String,
+    val createdTime: LocalDateTime
+)
