+++ src/main/kotlin/bingsoochef/bingsoochef/toppping/application/dto/ToppingInfo.kt
@@ -1,32 +1,31 @@
-package bingsoochef.bingsoochef.toppping.presentation.res
+package bingsoochef.bingsoochef.toppping.application.dto
 
 import bingsoochef.bingsoochef.toppping.domain.Topping
 import java.time.LocalDateTime
 
-data class ToppingDto (
+data class ToppingInfo (
     val toppingId: Long,
     val chefId: Long,
     val bingsooId: Long,
     val toppingTypeId: Long,
     val chefName: String,
-    val toppingTitle: String,
-    val toppingContent: String,
-    val toppingPosition: Long,
-    val toppingCreatedTime: LocalDateTime
+    val title: String,
+    val content: String,
+    val position: Long,
+    val createdTime: LocalDateTime
 ) {
-
     companion object {
-        fun from(topping : Topping): ToppingDto {
-            return ToppingDto(
+        fun from(topping: Topping): ToppingInfo {
+            return ToppingInfo(
                 toppingId = topping.id!!,
                 chefId = topping.chef.userId!!,
                 bingsooId = topping.bingsoo.id!!,
                 toppingTypeId = topping.toppingType.id!!,
                 chefName = topping.chefName,
-                toppingTitle = topping.title,
-                toppingContent = topping.content,
-                toppingPosition = topping.position,
-                toppingCreatedTime = topping.createdTime
+                title = topping.title,
+                content = topping.content,
+                position = topping.position,
+                createdTime = topping.createdTime
             )
         }
     }
